requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
idna==3.3
pushbullet.py==0.12.0
pycparser==2.21
python-magic==0.4.24
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.1